Remove the separate handler file - move to controller
For easier debugging. Now only the server
should be patched.
